UX-440 Abstract tab constructor into a hook
