+++ packages/matchbox/src/components/TimelineTabs/TimelineTabs.js
@@ -4,8 +4,8 @@ import styled from 'styled-components';
 import { margin } from 'styled-system';
 import { createPropTypes } from '@styled-system/prop-types';
 import { pick } from '../../helpers/props';
-import { onKey, onKeys } from '../../helpers/keyEvents';
 import Tab from './Tab';
+import useTabConstructor from './useTabConstructor';
 
 const Container = styled.div`
   ${margin}
@@ -16,88 +16,11 @@ const TimelineTabs = React.forwardRef(function TimelineTabs(props, userRef) {
   const systemProps = pick(rest, margin.propNames);
 
   const tabs = React.Children.toArray(children);
-  const tabRefs = React.useRef({ current: [] });
-  const focusRef = React.useRef();
-
-  const [selectedIndex, setSelectedIndex] = React.useState(initialIndex || 0);
-
-  // Preps array of tab item refs
-  React.useEffect(() => {
-    tabRefs.current = new Array(tabs.length);
-    return () => {
-      tabRefs.current = [];
-    };
-  }, [tabs, children]);
-
-  const onFocusContainerKeyDown = e => {
-    const isWithin = focusRef.current && focusRef.current.contains(e.currentTarget);
-
-    if (!isWithin) {
-      return;
-    }
-
-    onKey('arrowDown', () => {
-      e.preventDefault();
-      if (selectedIndex === tabs.length - 1) {
-        setSelectedIndex(0);
-      } else {
-        setSelectedIndex(selectedIndex + 1);
-      }
-    })(e);
-
-    onKey('arrowUp', () => {
-      e.preventDefault();
-      if (selectedIndex === 0) {
-        setSelectedIndex(tabs.length - 1);
-      } else {
-        setSelectedIndex(selectedIndex - 1);
-      }
-    })(e);
-
-    onKeys(['home', 'pageDown'], () => {
-      e.preventDefault();
-      setSelectedIndex(0);
-    })(e);
-
-    onKeys(['end', 'pageUp'], () => {
-      e.preventDefault();
-      setSelectedIndex(tabs.length - 1);
-    })(e);
-  };
-
-  // Focuses on tab items when focused index changes
-  // Optionally calls onSelect
-  React.useLayoutEffect(() => {
-    // Does not continue when focus is reset, when moving focus outside the container
-    if (focusRef.current && !focusRef.current.contains(document.activeElement)) {
-      return;
-    }
-
-    if (tabRefs.current[selectedIndex]) {
-      tabRefs.current[selectedIndex].focus();
-    }
-  }, [selectedIndex]);
-
-  const constructedItems = React.useMemo(() => {
-    return tabs.map((tab, i) => ({
-      ...tab,
-      ref: n => (tabRefs.current[i] = n),
-      props: {
-        ...tab.props,
-        handleParentSelect: () => setSelectedIndex(i),
-        selected: selectedIndex === i,
-      },
-    }));
-  }, [tabs, selectedIndex]);
+  const { tabs: constructedItems, focusContainerProps } = useTabConstructor({ initialIndex, tabs });
 
   return (
     <Container ref={userRef} data-id={rest['data-id']} id={id} {...systemProps}>
-      <div
-        role="tablist"
-        aria-orientation="vertical"
-        ref={focusRef}
-        onKeyDown={onFocusContainerKeyDown}
-      >
+      <div role="tablist" aria-orientation="vertical" {...focusContainerProps}>
         {constructedItems}
       </div>
     </Container>
